oops, `img_conv` didn't convert colors properly before!

+++ src/ColorsEtc.hpp
@@ -23,22 +23,22 @@ public:     // constants
 public:     // variables
     u16 data = 0u;
 public:     // functions
-    constexpr inline void set_r(u32 n_r) {
+    constexpr inline void set_r(u16 n_r) {
         set_bits_with_range(data, n_r, BIT_POS_HI_R, BIT_POS_LO_R);
     }
-    constexpr inline u32 get_r() {
+    constexpr inline u16 get_r() {
         return get_bits_with_range(data, BIT_POS_HI_R, BIT_POS_LO_R);
     }
-    constexpr inline void set_g(u32 n_g) {
+    constexpr inline void set_g(u16 n_g) {
         set_bits_with_range(data, n_g, BIT_POS_HI_G, BIT_POS_LO_G);
     }
-    constexpr inline u32 get_g() {
+    constexpr inline u16 get_g() {
         return get_bits_with_range(data, BIT_POS_HI_G, BIT_POS_LO_G);
     }
-    constexpr inline void set_b(u32 n_b) {
+    constexpr inline void set_b(u16 n_b) {
         set_bits_with_range(data, n_b, BIT_POS_HI_B, BIT_POS_LO_B);
     }
-    constexpr inline u32 get_b() {
+    constexpr inline u16 get_b() {
         return get_bits_with_range(data, BIT_POS_HI_B, BIT_POS_LO_B);
     }
